requirements.txt
bottle==0.12.25
mysql-connector-python~=8.3.0